+++ apps/www/content/docs/cn/markdown.mdx
@@ -89,7 +89,7 @@ const content = editor.api.markdown.serialize();
 MarkdownPlugin.configure({
   options: {
     remarkPlugins: [remarkMath, remarkGfm, remarkMdx],
-     nodes: {
+    rules: {
       date: {
         deserialize(mdastNode: MdMdxJsxTextElement, deco, options) {
           if (mdastNode.children?.[0] && 'value' in mdastNode.children[0]) {
@@ -117,7 +117,6 @@ MarkdownPlugin.configure({
         },
       },
     },
-    remarkPlugins: [remarkMath, remarkGfm, remarkMdx],
   },
 });
 ```
@@ -152,9 +151,9 @@ MarkdownPlugin.configure({
     这对于需要复杂逻辑来决定是否包含某个节点的场景很有用。
     默认值：`null`
   </APIItem>
-  <APIItem name="nodes" type="TNodes | null">
+  <APIItem name="rules" type="TRules | null">
     定义如何将 Markdown 语法元素转换为 Slate 编辑器元素的规则，或如何将 Slate 编辑器元素转换为 Markdown 语法元素的规则。包括段落、标题、列表、链接、图片等元素的转换规则。设置为 null 时将使用默认转换规则。
-    默认值：`undefined`
+    默认值：`null`
   </APIItem>
   <APIItem name="remarkPlugins" type="Plugin[]">
     用于扩展 Markdown 解析和序列化功能的 remark 插件数组。例如，可以添加 remark-gfm 来支持 GFM 语法，remark-math 来支持数学公式等。这些插件将在解析和生成 Markdown 文本时使用。
@@ -194,7 +193,7 @@ MarkdownPlugin.configure({
   <APIItem name="memoize" type="boolean" optional>
     启用带有 `_memo` 属性的块级记忆化，使其与 `PlateStatic` 记忆化兼容。
   </APIItem>
-  <APIItem name="nodes" type="TNodes | null" optional>
+  <APIItem name="rules" type="TRules" optional>
     定义如何将 Markdown 语法元素转换为 Slate 编辑器元素的规则。
   </APIItem>
   <APIItem name="parser" type="ParseMarkdownBlocksOptions" optional>
@@ -234,7 +233,7 @@ MarkdownPlugin.configure({
   <APIItem name="allowNode" type="AllowNodeConfig" optional>
     自定义节点过滤函数。在 allowedNodes/disallowedNodes 检查后调用。
   </APIItem>
-  <APIItem name="nodes" type="TNodes | null" optional>
+  <APIItem name="rules" type="TRules" optional>
     定义如何将 Slate 编辑器元素转换为 Markdown 语法元素的规则。
   </APIItem>
   <APIItem name="remarkPlugins" type="Plugin[]" optional>
