+++ packages/markdown/src/lib/deserializer/deserializeMd.ts
@@ -10,7 +10,7 @@ import type { TRules } from '../rules';
 
 import { mdastToSlate } from './mdastToSlate';
 import { type ParseMarkdownBlocksOptions, parseMarkdownBlocks } from './utils';
-import { getMergedOptionsDeserialize } from './utils/getMergedOptions';
+import { getMergedOptionsDeserialize } from './utils/getMergedOptionsDeserialize';
 
 // TODO: fixes tests
 
@@ -20,9 +20,9 @@ export type DeserializeMdOptions = {
   disallowedNodes?: NodesConfig;
   editor?: SlateEditor;
   memoize?: boolean;
-  nodes?: TRules | null;
   parser?: ParseMarkdownBlocksOptions;
   remarkPlugins?: Plugin[];
+  rules?: TRules | null;
   splitLineBreaks?: boolean;
 };
