+++ packages/markdown/src/lib/serializer/utils/getSerializerByKey.ts
@@ -6,7 +6,7 @@ export const getSerializerByKey = (
   key: string,
   options: SerializeMdOptions
 ) => {
-  const nodes = options.nodes;
+  const nodes = options.rules;
 
   return nodes?.[key]?.serialize === undefined
     ? defaultRules[key]?.serialize
